Класс вызва программы
